+++ internal/dao/filters.go
@@ -51,11 +51,11 @@ func (f OrderByFollowersFilter) Apply(db *gorm.DB) *gorm.DB {
 	return db.Order("followers_count desc")
 }
 
-type OrderByMembersFilter struct {
+type OrderByVotersFilter struct {
 }
 
-func (f OrderByMembersFilter) Apply(db *gorm.DB) *gorm.DB {
-	return db.Order("members_count desc")
+func (f OrderByVotersFilter) Apply(db *gorm.DB) *gorm.DB {
+	return db.Order("voters_count desc")
 }
 
 type DaoIDsFilter struct {