+++ internal/dao/models.go
@@ -105,7 +105,7 @@ type Dao struct {
 	Template       string
 	ParentID       *uuid.UUID
 	ActivitySince  int
-	MembersCount   int
+	VotersCount    int
 }
 
 func convertToCoreEvent(dao Dao) events.DaoPayload {